move the class VVHVLocalization from h file to cpp file as it is implementation details
